Add html boilerplate and basic block structure
This commit adds the html boilerplate and begins laying out the block
structure of the page. The page will have a large block at the top that
contains both the header and a hero section, a second block with some
information, a third block which presents a quote, a fourth block which
contains a call to action button, and a footer at the very bottom of
the page.