Begin adding header block content and styling
This commit adds the content to the header block, which contains the
header and the hero. The styling has been started, including defining
all the font properties, and some flex declarations have been added;
but the proportioning of the flex elements still needs some work,
especially in the header bar.